+++ common-core/src/main/java/cn/featherfly/common/lang/reflect/GenericClass.java
@@ -2,15 +2,16 @@
 package cn.featherfly.common.lang.reflect;
 
 import cn.featherfly.common.lang.AssertIllegalArgument;
-import cn.featherfly.common.lang.GenericType;
 
 /**
  * GenericClass .
  *
  * @author zhongj
  * @param <T> 类型
+ * @deprecated 后续删除，请使用{@link ClassType}替换
  */
-public class GenericClass<T> implements GenericType<T> {
+@Deprecated
+public class GenericClass<T> implements Type<T> {
 
     private Class<T> c;
 
@@ -61,4 +62,13 @@ public class GenericClass<T> implements GenericType<T> {
     public String getTypeName() {
         return c.getName();
     }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String toString() {
+        return this.getClass().getSimpleName() + " [type =" + getTypeName() + "]";
+    }
+
 }