+++ common-core/src/main/java/cn/featherfly/common/lang/reflect/MethodDescriptor.java
@@ -4,16 +4,17 @@ package cn.featherfly.common.lang.reflect;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 
-
 /**
  * <p>
  * MethodDescriptor
  * </p>
- * 
+ *
  * @author zhongj
+ * @deprecated
  */
+@Deprecated
 public class MethodDescriptor {
-    
+
     /**
      * @param method method
      */
@@ -27,7 +28,7 @@ public class MethodDescriptor {
             mp.setAnnotations(as);
             mp.setType(types[i]);
             mp.setMethod(method);
-//            mp.setName(name); 使用字节码技术才能获取
+            //            mp.setName(name); 使用字节码技术才能获取
             methodParameters[i] = mp;
         }
     }
@@ -38,6 +39,7 @@ public class MethodDescriptor {
 
     /**
      * 返回name
+     *
      * @return name
      */
     public String getName() {
@@ -46,6 +48,7 @@ public class MethodDescriptor {
 
     /**
      * 返回methodParameters
+     *
      * @return methodParameters
      */
     public MethodParameter[] getMethodParameters() {
@@ -54,15 +57,16 @@ public class MethodDescriptor {
 
     /**
      * 设置methodParameters
+     *
      * @param methodParameters methodParameters
      */
     public void setMethodParameters(MethodParameter[] methodParameters) {
         this.methodParameters = methodParameters;
     }
 
-    
     /**
      * 返回returnType
+     *
      * @return returnType
      */
     public Class<?> getReturnType() {