+++ common-core/src/main/java/cn/featherfly/common/lang/vt/ContainerType.java
@@ -1,0 +1,21 @@
+
+/*
+ * All rights Reserved, Designed By zhongj
+ * @Title: ContainerType.java
+ * @Package cn.featherfly.common.lang
+ * @Description: ContainerType
+ * @author: zhongj
+ * @date: 2022-11-10 14:11:10
+ * @Copyright: 2022 www.featherfly.cn Inc. All rights reserved.
+ */
+package cn.featherfly.common.lang.vt;
+
+/**
+ * ContainerType.
+ *
+ * @author zhongj
+ * @param <T> the generic type
+ */
+public interface ContainerType<T> extends ValueType<T> {
+
+}