+++ common-core/change-log.md
@@ -1,9 +1,13 @@
-# 1.10
+[ ] TODO 重构GenericType GenericClass BeanProperty这一些列的类，用于普通类，泛型类（Optional、Collection、Map）等
+
+# 1.10.8 
 
 1. WordUtils加入lowerCaseFirst方法
 2. 加入StringLinker类
 3. Strings加入trimStart(String,String),trimEnd(String,String),trimStartEnd(String,String),trimStartEndBlank(String)
-4. TODO 重构GenericType GenericClass BeanProperty这一些列的类，用于普通类，泛型类（Optional、Collection、Map）等
+4. cn.featherfly.common.enums包重构为cn.featherfly.common.operator并移动到common-api
+
+
 
 # 1.10.7 2022-07-15